+++ package.json
@@ -21,8 +21,8 @@
     "fs": "^0.0.1-security",
     "gray-matter": "^4.0.3",
     "js-cookie": "^3.0.1",
-    "next": "13.4.5-canary.6",
-    "next-auth": "^4.22.1",
+    "next": "13.4.6-canary.0",
+    "next-auth": "4.22.1",
     "next-mdx-remote": "^4.4.1",
     "react": "^18.2.0",
     "react-confetti": "^6.1.0",
